FIX: 비콘의 좌표 정보를 정수가 아닌 실수로 저장

+++ backend/jangan-backend/src/main/java/com/ssafy/jangan_backend/beacon/dto/response/ResponseExitBeaconDto.java
@@ -17,8 +17,8 @@ public class ResponseExitBeaconDto {
     private Integer mapId;
     private Integer beaconCode;
     private String name;
-    private Integer coordX;
-    private Integer coordY;
+    private Double coordX;
+    private Double coordY;
     private String cctvIp;
     @JsonProperty("is_cctv")
     private Boolean isCctv;
